Update downloader.py
Add option --only-flv so that just the flv can be downloaded.  This is useful if you got the subs in advance and edited them for later packing.  This also avoids the overwrite prompt if the sub already exists in the destination folder.

Note: In this edit the output still indicates that subs have been downloaded even though the download bit is skipped by faking that subs are hardcoded.  The same thing happens though even when subs are hardcoded for real or it couldn't find subs.